% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openness.R
\name{density_openness}
\alias{density_openness}
\alias{weighted_o}
\alias{weighted_inv_o}
\title{Density function of openness}
\usage{
density_openness(x, a, b, c, d, S = 1)

weighted_o(x, a, b, c, d, S = 1)

weighted_inv_o(x, a, b, c, d, S = 1)
}
\arguments{
\item{x}{numeric vector of quantiles.}

\item{a}{gamma shape parameter.}

\item{b}{gamma rate parameter.}

\item{c}{inverse-gamma shape parameter.}

\item{d}{inverse-gamma rate parameter.}

\item{S}{scaling constant}
}
\value{
\itemize{
\item \code{density_openness} gives scaled density.
\item \code{weighted_o} gives density-weighted openness.
\item \code{weighted_inv_o} gives density-weighted 1/openness.
}
}
\description{
Density function used for calculating expected values of openness and 1/openness.
The function is described by the product of gamma and inverse-gamma distributions.
}
