% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openness.R
\name{E_openness}
\alias{E_openness}
\title{Calculate expected values of openness and 1/openness}
\usage{
E_openness(lambda, chi, psi, alpha)
}
\arguments{
\item{lambda}{numeric vector of GIG parameter lambda.}

\item{chi}{numeric vector of GIG parameter chi. Must be positive.}

\item{psi}{numeric vector of GIG parameter psi. Must be positive.}

\item{alpha}{numeric vector of inverse-gamma shape parameter.}
}
\value{
A list object of the following expected values:
\describe{
\item{o}{numeric vector of expected opennes.}
\item{inv_o}{numeric vector of expected 1/openness.}
}
}
\description{
Expected values of openness and 1/openness using \code{stats::integrate} with
their density function described by the product of gamma and inverse-gamma distributions.
}
