% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openness.R
\name{exec_E_openness}
\alias{exec_E_openness}
\title{Execute E_openness}
\usage{
exec_E_openness(v)
}
\arguments{
\item{v}{numeric vector in the form \code{c(o, inv_o, lambda, chi, psi, alpha)}.
If both \code{o} and \code{inv_o} are valid, do nothing and return the \code{o} and \code{inv_o}.
If not, \code{E_openness} is executed with \code{lambda}, \code{chi}, \code{psi}, and \code{alpha} as parameters.}
}
\value{
A list object of the following expected values:
\describe{
\item{o}{expected opennes.}
\item{inv_o}{expected 1/openness.}
}
}
\description{
Wrapper for \code{E_openness} to execute through \code{apply} function family.
}
