% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_data}
\alias{load_data}
\title{Load data files}
\usage{
load_data(config, chr, region, rlen_unit = 1000)
}
\arguments{
\item{config}{a list object of configuration containing the following components:
\describe{
\item{contacts}{file of Enhancer-gene contacts.}
\item{hic_contact_colname}{contact frequency column name in enhancer-gene
contacts file.}
\item{window}{analyze enhancer-gene pairs with in this distance.}
\item{rnas}{file of gene expressions.}
\item{rna_id_colname}{gene id column name in gene expressions file.
Gene id must be unique in the file.}
\item{rna_length_colname}{transcript length column name in gene expressions
file.}
\item{rna_count_colname}{read count column name in gene expressions file.}
\item{rna_cutoff_colname}{column name of values examined for expression
cutoff in gene expressions file.}
\item{rna_cutoff_value}{expression cutoff. Genes whose exprssion greater
than this value will be analysed.}
\item{enhancers}{file of enhancer elements.}
\item{promoters}{file of genes promoter elements.}
\item{burst_sizes}{file of log10(genomic sequence-based burst size).}
}}

\item{chr}{chromosome to be analyzed}

\item{region}{region on chromosome to be analyzed}

\item{rlen_unit}{unit length used for transcript, enhancer and promoter length}
}
\value{
A list object of input data for the variational inference,
containing the following components:
\describe{
\item{chr}{chromosome.}
\item{region}{region on chromosome.}
\item{gene_name}{character vector of gene names.}
\item{gene_tss}{numeric vector of gene TSSs.}
\item{rna_count}{numeric vector of sequence read counts on genes.}
\item{rna_rlen}{numeric vector of relative effective lengths of genes.}
\item{pro_name}{character vector of gene promoter names.}
\item{pro_count}{numeric vecor of sequence read counts gene promoters.}
\item{pro_rlen}{numeric vecor of relative lengths of gene promoters.}
\item{enh_name}{character vector of enhancer names.}
\item{enh_count}{numeric vecor of sequence read counts enhancers.}
\item{enh_rlen}{numeric vecor of relative lengths of enhancers.}
\item{n_gene}{total number of genes.}
\item{n_enh}{total number of enhancers.}
\item{b_size}{numeric vector of genomic sequence-based burst size.}
\item{contact}{matrix (genes x enhancers) of contact frequencies.}
\item{sp_contact}{sparse matrix (genes x enhancers) of contact frequencies.}
\item{sp_distance}{sparse matrix (genes x enhancers) of genomic distances.}
}
}
\description{
Load data files according to configuration
}
