% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVEA.R
\name{make_initial_z}
\alias{make_initial_z}
\title{Make an initial resultant object}
\usage{
make_initial_z(ls_x, n_iter, no_le_enh)
}
\arguments{
\item{ls_x}{an input-data list object containing the following components:
\describe{
\item{chr}{chromosome.}
\item{region}{region on chromosome.}
\item{gene_name}{character vector of gene names.}
\item{gene_tss}{numeric vector of gene TSSs.}
\item{rna_count}{numeric vector of sequence read counts on genes.}
\item{rna_rlen}{numeric vector of relative effective lengths of genes.}
\item{pro_name}{character vector of gene promoter names.}
\item{pro_count}{numeric vecor of sequence read counts gene promoters.}
\item{pro_rlen}{numeric vecor of relative lengths of gene promoters.}
\item{enh_name}{character vector of enhancer names.}
\item{enh_count}{numeric vecor of sequence read counts enhancers.}
\item{enh_rlen}{numeric vecor of relative lengths of enhancers.}
\item{n_gene}{total number of genes.}
\item{n_enh}{total number of enhancers.}
\item{b_size}{numeric vector of genomic sequence-based burst size.}
\item{contact}{matrix (genes x enhancers) of contact frequencies.}
\item{sp_contact}{sparse matrix (genes x enhancers) of contact frequencies.}
\item{sp_distance}{sparse matrix (genes x enhancers) of genomic distances.}
}}

\item{n_iter}{Maximum number of iterations}

\item{no_le_enh}{Use IVEA_nolE model or not}
}
\value{
An initial resultant list object containing the following components:
\describe{
\item{k}{vector of estimated scaling factor values.}
\item{pro_openness}{matrix of estimated opennesses of gene promoters.}
\item{enh_openness}{matrix of estimated opennesses of enhancers.}
\item{pro_inv_openness}{matrix of estimated 1/opennesses of gene promoters.}
\item{enh_inv_openness}{matrix of estimated 1/opennesses of enhancers.}
\item{pro_activity}{matrix of estimated activities of gene promoters.}
\item{enh_activity}{matrix of estimated activities of enhancers.}
\item{pro_log_activity}{matrix of estimated log activities of gene promoters.}
\item{enh_log_activity}{matrix of estimated log activities of enhancers.}
}
}
\description{
Make a list object for latent variables to be estimated and set the initial values
}
