% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVEA.R
\name{run_variational_bayes}
\alias{run_variational_bayes}
\title{Run variational Bayesian inference}
\usage{
run_variational_bayes(param, ls_x)
}
\arguments{
\item{param}{a parameter list object containing the following components:
\describe{
\item{alpha_pro}{Hyperparameter alpha for the gamma prior of promoter activity.}
\item{alpha_enh}{Hyperparameter alpha for the gamma prior of enhancer activity.}
\item{alpha_o}{Hyperparameter alpha for the gamma prior of openness.}
\item{beta_o}{Hyperparameter beta for the gamma prior of openness.}
\item{alpha_k}{Hyperparameter alpha for the gamma prior of skaler k.}
\item{beta_k}{Hyperparameter beta for the gamma prior of skaler k.}
\item{no_le_enh}{use IVEA_nolE model or not.}
\item{max_iteration}{Maximum iterations in variational inference.}
\item{stop_criterion}{Stop iteration when all relative change of
enhancer activities are less than this value.}
}}

\item{ls_x}{an input-data list object containing the following components:
\describe{
\item{chr}{chromosome.}
\item{region}{region on chromosome.}
\item{gene_name}{character vector of gene names.}
\item{gene_tss}{numeric vector of gene TSSs.}
\item{rna_count}{numeric vector of sequence read counts on genes.}
\item{rna_rlen}{numeric vector of relative effective lengths of genes.}
\item{pro_name}{character vector of gene promoter names.}
\item{pro_count}{numeric vecor of sequence read counts gene promoters.}
\item{pro_rlen}{numeric vecor of relative lengths of gene promoters.}
\item{enh_name}{character vector of enhancer names.}
\item{enh_count}{numeric vecor of sequence read counts enhancers.}
\item{enh_rlen}{numeric vecor of relative lengths of enhancers.}
\item{n_gene}{total number of genes.}
\item{n_enh}{total number of enhancers.}
\item{b_size}{numeric vector of genomic sequence-based burst size.}
\item{contact}{matrix (genes x enhancers) of contact frequencies.}
\item{sp_contact}{sparse matrix (genes x enhancers) of contact frequencies.}
\item{sp_distance}{sparse matrix (genes x enhancers) of genomic distances.}
}}
}
\value{
A resultant list object containing the following components:
\describe{
\item{k}{vector of estimated scaling factor values.}
\item{pro_openness}{matrix of estimated opennesses of gene promoters.}
\item{enh_openness}{matrix of estimated opennesses of enhancers.}
\item{pro_inv_openness}{matrix of estimated 1/opennesses of gene promoters.}
\item{enh_inv_openness}{matrix of estimated 1/opennesses of enhancers.}
\item{pro_activity}{matrix of estimated activities of gene promoters.}
\item{enh_activity}{matrix of estimated activities of enhancers.}
\item{pro_log_activity}{matrix of estimated log activities of gene promoters.}
\item{enh_log_activity}{matrix of estimated log activities of enhancers.}
}
}
\description{
Perform variational inference of regulatory elements (promoter and enhancers)
opennesses and activities.
}
\details{
The parameters and expectations of the latent variables are iteratively updated
using the equations described in (reference) until convergence of the enhancer
activities judged by \code{stop_criterion} or \code{max_iteration}.
}
