% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openness.R
\name{E_openness}
\alias{E_openness}
\title{Calculate expected values of openness and 1/openness}
\usage{
E_openness(a, b, c, d)
}
\arguments{
\item{a}{gamma shape parameter.}

\item{b}{gamma rate parameter.}

\item{c}{inverse-gamma shape parameter.}

\item{d}{inverse-gamma rate parameter.}
}
\value{
A list object of the following expected values:
\describe{
\item{o}{numeric vector of expected opennes.}
\item{inv_o}{numeric vector of expected 1/openness.}
}
}
\description{
Expected values of openness and 1/openness using \code{stats::integrate} with
their density function described by the product of gamma and inverse-gamma distributions.
}
