% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openness.R
\name{exec_Egig_x}
\alias{exec_Egig_x}
\alias{exec_Egig_inv_x}
\title{Execute ghyp::Egig}
\usage{
exec_Egig_x(v)

exec_Egig_inv_x(v)
}
\arguments{
\item{v}{numeric vector of parameters for GIG in the form \code{c(lambda, chi, psi)}.}
}
\value{
Expected \code{x} or \code{1/x} from ghyp::Egig
\itemize{
\item \code{exec_Egig_x} gives expected \code{x} from ghyp::Egig
\item \code{exec_Egig_inv_x} gives expected \code{1/x} from ghyp::Egig
}
}
\description{
Wrapper for \code{ghyp::Egig} to execute through \code{apply} function family.
}
