% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openness.R
\name{get_openness}
\alias{get_openness}
\title{Get expected openness and 1/openness}
\usage{
get_openness(lambda, chi, psi, alpha)
}
\arguments{
\item{lambda}{numeric vector of GIG parameter lambda.}

\item{chi}{numeric vector of GIG parameter chi. Must be positive.}

\item{psi}{numeric vector of GIG parameter psi. Must be positive.}

\item{alpha}{numeric vector of inverse-gamma shape parameter.}
}
\value{
A list object of the following expected values:
\describe{
\item{o}{numeric vector of expected opennes.}
\item{inv_o}{numeric vector of expected 1/openness.}
}
}
\description{
Get expected openness and 1/openness through \code{ghyp::Egig} and \code{E_openness} functions.
}
\details{
\code{ghyp::Egig} and \code{E_openness} work for different sets of parameters.
To reduce failures in obtaining expected values,
first execute \code{ghyp::Egig}, next perform \code{E_openness} if \code{ghyp::Egig} failed.
}
